+++ packages/webawesome/scripts/build.js
@@ -399,7 +399,7 @@ export async function build(options = {}) {
         return async filename => {
           const changedFile = relative(getRootDir(), filename);
 
-          if (evt === 'changed') {
+          if (evt === 'change') {
             spinner.info(`File modified ${chalk.gray(`(${changedFile})`)}`);
           } else if (evt === 'unlink') {
             spinner.info(`File deleted ${chalk.gray(`(${changedFile})`)}`);
@@ -471,7 +471,7 @@ export async function build(options = {}) {
           }
           await generateDocs({ spinner });
 
-          if (typeof options.beforeWatchEvent === 'function') {
+          if (typeof options.afterWatchEvent === 'function') {
             await options.afterWatchEvent(evt, filename);
           }
           reload();